fix: clear Sonar new-code issues for seat launcher PR
Coverage already passed at 84.7%; the quality gate was failing on 33 new
code smells (docstrings, duplicated literals, file length, test shape).
Refactor lifecycle helpers, move kiosk spawn and overlay cleanup out, and
tighten VM and observation contracts so assert_no_new_issues passes.
