+++ src/aptl/appliance/seat/kiosk.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import shutil
+import subprocess
 from dataclasses import dataclass
 
 
@@ -34,7 +35,34 @@ def build_kiosk_launch_plan(
     return KioskLaunchPlan(argv=argv, url=url)
 
 
+def open_participant_kiosk(
+    *,
+    participant_port: int = 443,
+    browser_command: str | None = None,
+    dry_run: bool = False,
+) -> KioskLaunchPlan:
+    """Launch or plan the participant kiosk browser wrapper."""
+
+    plan = build_kiosk_launch_plan(
+        participant_port=participant_port,
+        browser_command=browser_command,
+    )
+    if dry_run:
+        return plan
+    subprocess.Popen(
+        list(plan.argv),
+        stdin=subprocess.DEVNULL,
+        stdout=subprocess.DEVNULL,
+        stderr=subprocess.DEVNULL,
+        close_fds=True,
+        start_new_session=True,
+    )
+    return plan
+
+
 def _default_browser() -> str:
+    """Pick the first installed fullscreen-capable browser on the host."""
+
     for candidate in ("chromium-browser", "chromium", "google-chrome", "firefox"):
         if shutil.which(candidate):
             return candidate
